requirements.txt
# Core dependencies for Vask Voice-Based AI Companion

# Speech Recognition
openai-whisper==20231117

# Language Model
llama-cpp-python==0.2.36

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.8

# Text-to-Speech
piper-tts==1.2.0

# Encryption
cryptography==41.0.7

# Configuration
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
hypothesis==6.92.1

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
